perf(render): skip clean frames; animate overlays without recompositing
needs_recomposite was set by every mutation path (via mark_layer_dirty
and ~30 direct sites) but never read — render() composited every layer
unconditionally on each dirty rAF tick, including cursor moves and
marching-ants animation. Wire the flag as a frame gate in render(), and
extract renderOverlayFrame() so ant/text-cursor animation repaints only
the 2D overlay canvas instead of forcing a full GPU composite at 60fps.

Delete the dirty_layers HashSet (written everywhere, read nowhere) and
the markLayerDirty export (no JS callers). The dev readPixels helper
now forces a recomposite explicitly since render() can skip.
