+++ src/engine-wasm/wasm-bridge.ts
@@ -84,7 +84,6 @@ import init, {
   renderShapeExpanded,
   uploadPathTexture,
   render,
-  markLayerDirty,
   markAllDirty,
   setImageExposure,
   setImageContrast,
@@ -389,7 +388,6 @@ export {
   renderShapeExpanded,
   uploadPathTexture,
   render,
-  markLayerDirty,
   markAllDirty,
   setImageExposure,
   setImageContrast,
